Unbind server start with renamed route
