+++ server.js
@@ -1,0 +1,6 @@
+const app = require('./src/routes');
+const port = 3000;
+
+app.listen(port, () => {
+  console.log(`listening at http://localhost:${port}`);
+});
