--- src/app.js
@@ -1,14 +1,0 @@
-const express = require('express');
-const app = express();
-// const port = 3000;
-
-app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
-
-/*
-app.listen(port, () => {
-  console.log(`listening at http://localhost:${port}`);
-});
-*/
-module.exports = app;
