+++ src/routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const app = express();
+
+app.get('/bars', (req, res) => {
+  res.send('Hello World!');
+});
+
+module.exports = app;
