+++ tests/routes.test.js
@@ -1,9 +1,9 @@
 const request = require('supertest');
-const app = require('../src/app');
+const app = require('../src/routes');
 
-describe('Root route', () => {
+describe('/bars', () => {
   it('returns 200 response', async () => {
-    const res = await request(app).get('/');
+    const res = await request(app).get('/bars');
 
     expect(res.statusCode).toEqual(200);
   });
